pyproject.toml
[project]
name = "photocampus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-graphql-jwt>=0.4.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.9",
    "graphene-django>=3.2.2",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.9",
]
